move story component to stories component
