+++ src/app/components/stories/stories.component.css
@@ -1,0 +1,12 @@
+:host {
+  width: 100%;
+}
+
+.thumbnail img {
+  max-width: 40px;
+  max-height: 40px;
+}
+
+.error {
+  color: #f44336;
+}
