--- src/app/components/story/story.component.css
@@ -1,8 +1,0 @@
-.thumbnail img {
-  max-width: 40px;
-  max-height: 40px;
-}
-
-.error {
-  color: #f44336;
-}
